requirements.txt
numpy>=1.17.0
torch==1.9.0+cu111
torchvision==0.10.0+cu111
Pillow>=6.1.0
scipy>=1.3.0
statsmodels>=0.10.1
pandas>=0.25.0
matplotlib
seaborn
tqdm
setGPU
pyyaml==5.4.1
EasyDict
